+++ include/UECS/CmptLocator.h
@@ -12,6 +12,7 @@ namespace Ubpa::UECS {
 	class CmptLocator {
 	public:
 		CmptLocator(Span<const CmptAccessType> types);
+		CmptLocator(CmptAccessTypeSet types);
 
 		CmptLocator();