+++ include/UECS/EntityMngr.h
@@ -36,6 +36,7 @@ namespace Ubpa::UECS {
 
 		// use RTDCmptTraits
 		Entity Create(Span<const CmptType> types);
+		Entity Create(CmptType type) { return Create({ &type, 1 }); }
 
 		Entity Instantiate(Entity);
 
@@ -46,25 +47,27 @@ namespace Ubpa::UECS {
 
 		// use RTDCmptTraits
 		void Attach(Entity, Span<const CmptType> types);
+		void Attach(Entity e, CmptType type) { Attach(e, { &type, 1 }); }
 
 		template<typename Cmpt, typename... Args>
 		Cmpt* Emplace(Entity, Args&&...);
 
+		void Detach(Entity, Span<const CmptType> types);
+		void Detach(Entity e, CmptType type) { Detach(e, { &type, 1 }); }
 		// use Detach(Entity, const CmptType*, size_t)
 		template<typename... Cmpts>
 		void Detach(Entity);
-		void Detach(Entity, Span<const CmptType> types);
 
+		bool Have(Entity, CmptType) const;
 		// use Have(Entity, CmptType)
 		template<typename Cmpt>
 		bool Have(Entity) const;
-		bool Have(Entity, CmptType) const;
 
-		// nullptr if not containts <Cmpt>
-		template<typename Cmpt>
-		Cmpt* Get(Entity) const;
 		// nullptr if not containts CmptType
 		CmptPtr Get(Entity, CmptType) const;
+		// use Get(Entity, CmptType)
+		template<typename Cmpt>
+		Cmpt* Get(Entity) const;
 
 		std::vector<CmptPtr> Components(Entity) const;
 
@@ -77,8 +80,6 @@ namespace Ubpa::UECS {
 		// use entry in reverse
 		const std::vector<size_t>& GetEntityFreeEntries() const noexcept { return entityTableFreeEntry; }
 		size_t GetEntityVersion(size_t idx) const noexcept { return entityTable.at(idx).version; }
-
-		std::tuple<bool, std::vector<CmptAccessPtr>> LocateSingletons(const SingletonLocator&) const;
 
 		bool IsSingleton(CmptType) const;
 		Entity GetSingletonEntity(CmptType) const;
@@ -110,6 +111,8 @@ namespace Ubpa::UECS {
 		template<typename... Cmpts>
 		Archetype* AttachWithoutInit(Entity);
 		Archetype* AttachWithoutInit(Entity, Span<const CmptType> types);
+
+		std::vector<CmptAccessPtr> LocateSingletons(const SingletonLocator&) const;
 
 		const std::set<Archetype*>& QueryArchetypes(const EntityQuery&) const;
 		mutable std::unordered_map<EntityQuery, std::set<Archetype*>> queryCache;