+++ src/core/CmptLocator.cpp
@@ -4,6 +4,12 @@
 
 using namespace Ubpa::UECS;
 
+CmptLocator::CmptLocator(CmptAccessTypeSet cmptTypes)
+	: cmptTypes{ std::move(cmptTypes) }
+{
+	UpdateHashCode();
+}
+
 CmptLocator::CmptLocator(Span<const CmptAccessType> types) {
 	for (const auto& type : types)
 		cmptTypes.insert(type);